glib: manually implement file_open_tmp
To make use of RawFd types
