+++ glib/Gir.toml
@@ -108,6 +108,9 @@ status = "generate"
     pattern = "io_.*"
     ignore = true # they are supposed to be used with IOChannel, which doesn't have safe bindings
     [[object.function]]
+    name = "file_open_tmp"
+    manual = true # Return a RawFd instead of a i32
+    [[object.function]]
     pattern = "atomic_.+"
     ignore = true
     [[object.function]]
